Add files via upload
These are two concept layout ideas that make it easy for a user to easily navigate between functions to pull their desired data.
One is a user interface that needs a bit more work to make it mobile responsive.

But the other one is perfectly mobile responsive. just type and enter.

Thinking about adding a light and dark theme template next so any project we want to add a light and dark theme on we can just pull the stuff from the folder here.